Upload print_duck.py
This is the first homework assignment for DAT-129. For this assignment, I drew a duck in a 10x10 grid by coloring certain squares within the grid and leaving specific squares blank. The picture was then converted into a binary string, where blank spaces equal 0 and colored in spaces equal 1. This string was coded into lists representing each line and housed within one larger list.
The program then takes the binary list and changes the values to so that 0 transforms into a blank space and one becomes @. Then there are a series of functions that rotate and flip the image. I wasn't able to figure out scaling the image in time, but hope that future iterations of this program will be able to do just that.